// src/app/pages/store/products/detail/interfaces/i-product-detail.ts
export interface IProductDetail
{
  productId: number;
  storeId: number;
  refModelId: number;
  refBrandId: number;
  productDesc: string;
  dateCreated: string;
  dateUpdated: string;
  productStatus: number;
  orderIndex: number;
  barcode: string;
  hasStock: boolean;
  sku: string;
  productNo: string;
  unitsWeight: number;
  boxSize: number;
  productType: string;
  keywords: string;
  volume: number;
  refCateIds: number[];
  vendorId: number;
  vendorName: string;
  vendorProductName: string;
  rrp: number;
  srp: number;
  remark: string;
  sellingShipping: number;
  refFlagIds: number[];
  refVariantIds: number[];
  buyingPrice: number;
  buyingRefCurrencyId: number;
  rate: number;
  link: number;
  uom: number;
  reorderQty: number;
  warehouseLoc: string;
  msrpUSD: number;
  competitorTHB: number;
  shippingCost: number;
  image: string;
  unitOnTheWay: number;
  unitsOnDesposit: number;
  unitsInWarehouse: number;
  unitsOnDisplay: number;
  unitsOnDemo: number;
  unitsOnQc: number;
  stockQty: number;
  url: string;
  uidTypeId: number;
}